+++ app/Http/Controllers/UsersController.php
@@ -7,27 +7,19 @@ use Illuminate\Foundation\Validation\ValidatesRequests;
 use JobApis\JobsToMail\Http\Requests\CreateUser;
 use JobApis\JobsToMail\Jobs\ConfirmUser;
 use JobApis\JobsToMail\Jobs\CreateUserAndSearch;
-use JobApis\JobsToMail\Jobs\UnsubscribeUser;
-use JobApis\JobsToMail\Repositories\Contracts\UserRepositoryInterface;
+use JobApis\JobsToMail\Jobs\DeleteUser;
+use JobApis\JobsToMail\Jobs\GetUserSearches;
 
 class UsersController extends BaseController
 {
     use DispatchesJobs, ValidatesRequests;
 
     /**
-     * UsersController constructor.
-     */
-    public function __construct(UserRepositoryInterface $users)
-    {
-        $this->users = $users;
-    }
-
-    /**
      * Home page and signup form
      */
     public function index()
     {
-        return view('users.welcome');
+        return view('users.index');
     }
 
     /**
@@ -61,9 +53,23 @@ class UsersController extends BaseController
      */
     public function unsubscribe(Request $request, $userId)
     {
-        $message = $this->dispatchNow(new UnsubscribeUser($userId));
+        $message = $this->dispatchNow(new DeleteUser($userId));
 
         $request->session()->flash($message->type, $message->message);
+
+        return redirect('/');
+    }
+
+    /**
+     * View searches for this user
+     */
+    public function searches(Request $request, $userId)
+    {
+        $results = $this->dispatchNow(new GetUserSearches($userId));
+
+        if (!$results->isEmpty()) {
+            return view('searches.index', ['searches' => $results]);
+        }
 
         return redirect('/');
     }