+++ app/Repositories/Contracts/SearchRepositoryInterface.php
@@ -4,5 +4,9 @@ interface SearchRepositoryInterface
 {
     public function create($userId = null, $data = []);
 
+    public function delete($id = null);
+
     public function getActive($userEmail = null);
+
+    public function getByUserId($userId = null);
 }