+++ app/Repositories/SearchRepository.php
@@ -35,6 +35,18 @@ class SearchRepository implements Contracts\SearchRepositoryInterface
     }
 
     /**
+     * Delete a search by ID
+     *
+     * @param string $id
+     *
+     * @return boolean
+     */
+    public function delete($id = null)
+    {
+        return $this->searches->where('id', $id)->delete();
+    }
+
+    /**
      * Get all active searches from confirmed users
      *
      * @param null | string $userEmail
@@ -48,5 +60,17 @@ class SearchRepository implements Contracts\SearchRepositoryInterface
             $query = $query->whereUserEmail($userEmail);
         }
         return $query->get();
+    }
+
+    /**
+     * Get searches by User Id
+     *
+     * @param string $userId
+     *
+     * @return \Illuminate\Database\Eloquent\Collection
+     */
+    public function getByUserId($userId = null)
+    {
+        return $this->searches->whereUserId($userId)->get();
     }
 }