+++ database/seeds/DatabaseTruncater.php
@@ -11,6 +11,7 @@ class DatabaseTruncater extends Seeder
         DB::statement('ALTER TABLE tokens DISABLE TRIGGER ALL;');
         DB::statement('TRUNCATE searches CASCADE');
         DB::statement('TRUNCATE tokens CASCADE');
+        DB::statement('TRUNCATE notifications CASCADE');
         DB::statement('TRUNCATE users CASCADE');
         DB::statement('ALTER TABLE tokens ENABLE TRIGGER ALL;');
         DB::statement('COMMIT;');