+++ resources/views/layouts/app.blade.php
@@ -3,9 +3,16 @@
     @include('layouts.head')
     <body>
         @include('layouts.flash-messages')
+
         <div class="container-fluid">
-            @yield('content')
+
+            <div class="main">
+                @yield('content')
+            </div>
+
         </div>
+
         @include('layouts.footer')
+
     </body>
 </html>